requirements.txt
numpy
opencv-contrib-python
pyrealsense2
PyYAML
scipy
git+https://github.com/kvasios/franky-remote.git

